+++ Makefile
@@ -1,10 +1,10 @@
 default: dev
 
 install:
-	@rm -f $HOME/go/bin/odas && go generate . > ./internal/commit.txt &&  go install .
+	@rm -f $HOME/go/bin/odas && go generate . > commit.txt &&  go install .
 build:
-	@go generate . > ./internal/commit.txt
+	@go generate . > commit.txt
 	@CGO_ENABLED=0 GOOS=linux go build -a -o ./bin/odas .
 dev:
-	@go generate . > ./internal/commit.txt
+	@go generate . > commit.txt
 	@go build -o ./bin/odas .
