--- internal/cmdBackupBackup.go
@@ -1,36 +1,0 @@
-/*
-Copyright © 2024 Peter Preeper
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program. If not, see <http://www.gnu.org/licenses/>.
-*/
-package internal
-
-import (
-	"github.com/spf13/cobra"
-)
-
-// backupCmd represents the backup command
-var backupCmd = &cobra.Command{
-	Use:     "backup",
-	Short:   "Backup database filestore and addons",
-	Long:    `Backup database filestore and addons`,
-	GroupID: "backup",
-	Run: func(cmd *cobra.Command, args []string) {
-		AdminBackup()
-	},
-}
-
-func init() {
-	rootCmd.AddCommand(backupCmd)
-}
