--- internal/cmdConfigHosts.go
@@ -1,53 +1,0 @@
-/*
-Copyright © 2024 Peter Preeper
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program. If not, see <http://www.gnu.org/licenses/>.
-*/
-package internal
-
-import (
-	"github.com/spf13/cobra"
-)
-
-// hostsCmd represents the hosts command
-var hostsCmd = &cobra.Command{
-	Use:     "hosts",
-	Short:   "update hosts file",
-	Long:    `update hosts file`,
-	GroupID: "config",
-	Run: func(cmd *cobra.Command, args []string) {
-		domain := ""
-		modlen := len(args)
-		if modlen == 0 {
-			domain = "local"
-		} else {
-			domain = args[0]
-		}
-		HostsUpdate(domain)
-	},
-}
-
-func init() {
-	rootCmd.AddCommand(hostsCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// hostsCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// hostsCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-}
