+++ templates/update
@@ -1,0 +1,2 @@
+#!/bin/bash
+sudo bash -c "apt update -y && apt full-upgrade -y && apt autoremove -y && apt autoclean -y"
